Use notes and many other tweaks

+++ src/main/kotlin/com/petertackage/geneticsound/FitnessFunction.kt
@@ -2,5 +2,5 @@ package com.petertackage.geneticsound
 
 interface FitnessFunction {
 
-    fun compare(target: ShortArray, proposed: ShortArray): Double
+    fun compare(target: ShortArray, proposed: ShortArray): Long
 }
